refactored react components and created splash page
