+++ react-app/src/components/Login/LoginForm.js
@@ -10,6 +10,7 @@ export default function LoginForm() {
     const [errors, setErrors] = useState([]);
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
+
     const user = useSelector((state) => state.session.user);
     const dispatch = useDispatch();
