+++ react-app/src/components/Nav/NavBar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import LogoutButton from "./auth/LogoutButton";
+import LogoutButton from "../Logout";
 
 import styles from "./NavBar.module.css";
 // className={styles. }
