+++ react-app/src/components/SplashPage/SplashPage.module.css
@@ -1,0 +1,3 @@
+.splash {
+    display: flex;
+}
