+++ react-app/src/components/SplashPage/index.js
@@ -1,0 +1,3 @@
+import SplashPage from "./SplashPage";
+
+export default SplashPage;
